# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TernakPro API",
        "description": "API untuk aplikasi Ternak yang menyediakan berbagai fitur terkait ternak dan manajemen pengguna.",
        "contact": {
            "email": "dopidev@gmail.com"
        },
        "version": "v1"
    },
    "paths": {
        "/api/v1/asset/{user_id}": {
            "get": {
                "tags": [
                    "Asset"
                ],
                "summary": "Mengambil asset berdasarkan user_id",
                "description": "Endpoint ini digunakan untuk mengambil semua data asset berdasarkan ID pengguna.",
                "operationId": "88c57d6dd096ab0e0fa5a297a616db16",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "ID pengguna",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data asset",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/MAsset"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Asset"
                ],
                "summary": "Mengupdate data asset",
                "description": "Endpoint ini digunakan untuk mengupdate data asset berdasarkan ID pengguna.",
                "operationId": "f96a74ff7a18ede6def1653cfa616724",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "ID pengguna",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama": {
                                        "type": "string",
                                        "example": "Sapi Holstein"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Asset berhasil diupdate",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MAsset"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Asset tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Asset tidak ditemukan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validasi gagal"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Asset"
                ],
                "summary": "Menyimpan asset baru",
                "description": "Endpoint ini digunakan untuk membuat data asset baru untuk pengguna tertentu.",
                "operationId": "eb72172ad715d04d11d952878f04d79f",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "ID pengguna",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama"
                                ],
                                "properties": {
                                    "nama": {
                                        "type": "string",
                                        "example": "Sapi Holstein"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Asset berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MAsset"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validasi gagal"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Asset"
                ],
                "summary": "Menghapus data asset",
                "description": "Endpoint ini digunakan untuk menghapus data asset berdasarkan ID pengguna.",
                "operationId": "5c6cad1b888562eaaebc188ba948c784",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "ID pengguna",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Asset berhasil dihapus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Asset berhasil dihapus"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Asset tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Asset tidak ditemukan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "MAsset": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "string",
                        "example": "user123"
                    },
                    "nama": {
                        "type": "string",
                        "example": "Sapi Holstein"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
...
            "Role": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nama_role": {
                        "type": "string",
                        "example": "Admin"
                    },
                    "is_aktif": {
                        "type": "boolean",
                        "example": true
                    },
                    "users": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/User"
                        }
                    }
                },
                "type": "object"
            },
...
            "User": {
                "properties": {
                    "uid": {
                        "type": "string",
                        "example": "user123"
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john.doe@example.com"
                    },
                    "role_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "no_telepon": {
                        "type": "string",
                        "example": "+6281234567890"
                    },
                    "profile_image": {
                        "type": "string",
                        "example": "/images/profile.jpg"
                    },
                    "email_verified_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-20T12:00:00Z"
                    },
                    "role": {
                        "$ref": "#/components/schemas/Role"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Masukkan token dalam format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}